decimais

+++ cálculo.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 int main() {
-    int lateral = 10;
-    int cquarto = 7;
-    int areaq;
-    int areas;
-    int areat;
-// int / i e a variavel dos numeros inteiros então o resulta como ex não pode dar 11/2 = 5.5
+    float lateral = 10;
+    float cquarto = 7;
+    float areaq;
+    float areas;
+    float areat;
+// float e uma variavel que serve para mostrar os decimais
     printf("progama para cálculo da área da casa\n");
     areas = lateral*lateral ;
-    printf("A área da sala é %i\n" , areas);
+    printf("A área da sala é %f\n" , areas);
     areaq = cquarto*(lateral/2);
-    printf("A área do quarto é %i\n" , areaq);
-    printf("A área do banheiro é %i\n" , areaq);
+    printf("A área do quarto é %f\n" , areaq);
+    printf("A área do banheiro é %f\n" , areaq);
     areat = areas + 2*areaq;
-    printf("A área total é %i\n" , areat);
+    printf("A área total é %f\n" , areat);
     return 0;
 }
